refactor: rename file "resolve-config" to "resolveConfig"
- rename file "resolve-config" to "resolveConfig"
- rename file "resolve-config.test" to "resolveConfig.test"

(rename to match other file names [camelCase])
