+++ packages/mongodb-memory-server-core/src/util/resolveConfig.ts
