+++ packages/mongodb-memory-server-global-4.4/index.js
@@ -2,7 +2,7 @@
 'use strict';
 Object.defineProperty(exports, '__esModule', { value: true });
 
-var setDefaultValue = require('../mongodb-memory-server-core/lib/util/resolve-config')
+var setDefaultValue = require('../mongodb-memory-server-core/lib/util/resolveConfig')
   .setDefaultValue;
 var version = require('./package.json').mongodb_version;
 setDefaultValue('VERSION', version);
